pyproject.toml
[tool.poetry]
name = "fashion-img-search"
version = "0.1.0"
description = ""
authors = []
license = "MIT"
readme = "README.md"
packages = [{ include = "fis" }]

[tool.poetry.dependencies]
python = "^3.7.1"
torch = "^1.12.1"
torchvision = "^0.13.1"
timm = "^0.6.7"
pandas = "1.3"
Pillow = "^9.2.0"
pymilvus = "^2.1.2"
python-dotenv = "^0.21.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
isort = "^5.10.1"
flake8 = "^5.0.4"
pre-commit = "^2.20.0"
mypy = "^0.971"
jupyterlab = "^3.4.7"

[tool.black]
line-length = 120
target-version = ["py37"]

[tool.mypy]
ignore_missing_imports = true
exclude = [""]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
